backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "argon2-cffi>=25.1.0",
    "fastapi[standard]>=0.135.3",
    "motor>=3.7.1",
    "passlib>=1.7.4",
    "pydantic-settings>=2.13.1",
    "pydantic[email]>=2.12.5",
    "pymongo>=4.16.0",
    "python-jose>=3.5.0",
]

[dependency-groups]
dev = [
    "black>=26.3.1",
]

# ------------------------
# BLACK (formatter)
# ------------------------
[tool.black]
line-length = 100
target-version = ["py310"]
skip-string-normalization = true

exclude = '''
/(
    \.venv
  | __pycache__
  | migrations
  | build
  | dist
)/
'''

# ------------------------
# RUFF (linter)
# ------------------------
[tool.ruff]
line-length = 100
select = ["E", "F", "I"]  # Errors, formatting, imports
ignore = []
fix = true

# ------------------------
# ISORT (import sorting)
# ------------------------
[tool.isort]
profile = "black"
line_length = 100
